Update README.md
One static simple chatbot which I made using python Dictionary. It's character sensitive so be careful with your choice.
Matrix Addition and Transposing function in the second file.

Number guessing game is there on the third. Which you are asked to provide two numbers between which a bot will randomly select one number and you will be asked to guess that particular number before some pre-determined steps which you will be asked to provide. The bot will notify you if you entered greater value or larger one, this way you will understand in between which value your desired value placed. so keep guessing till either you reached your maximum guessing or you already guessed that value.

Fourth one is Fizz-buzz game, as many of you know it. In that function you  need to specify only two integer values upper and lower bound. And between them if one value is divisible by 3 and 5 both, the number will be replaced by "Fizz-buzz" and if one value is divisible by only 3 but not five, it'll be replaced by "Fizz" and by only 5 but not by 3, then it will be replaced by buzz.